addressing more feedback
